Move states/channels/index.ts to channel/state.ts
