+++ packages/wallet/src/__stories__/index.tsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import WalletContainer from '../containers/Wallet';
 import { Provider } from 'react-redux';
 import * as walletStates from '../redux/state';
-import * as channelStates from '../redux/states/channels';
+import * as channelStates from '../redux/channel/state';
 import * as fundingStates from '../redux/states/channels/funding/index';
 import '../index.scss';
 import * as scenarios from '../redux/reducers/__tests__/test-scenarios';
