+++ packages/wallet/src/redux/reducers/__tests__/closing.test.ts
@@ -1,6 +1,6 @@
 import { closingReducer } from '../channels/closing';
 
-import * as states from '../../states/channels';
+import * as states from '../../channel/state';
 import * as actions from '../../actions';
 import * as outgoing from 'magmo-wallet-client/lib/wallet-events';
 import * as scenarios from './test-scenarios';
