+++ packages/wallet/src/redux/reducers/__tests__/responding.test.ts
@@ -1,6 +1,6 @@
 import { respondingReducer } from '../channels/responding';
 
-import * as states from '../../states/channels';
+import * as states from '../../channel/state';
 import * as actions from '../../actions';
 
 import { itTransitionsToChannelStateType, itDoesntTransition } from './helpers';
