+++ packages/wallet/src/redux/reducers/channels/withdrawing.ts
@@ -1,4 +1,4 @@
-import * as states from '../../states/channels';
+import * as states from '../../channel/state';
 import * as actions from '../../actions';
 import { unreachable } from '../../../utils/reducer-utils';
 import { handleSignatureAndValidationMessages } from '../../../utils/state-utils';
