+++ packages/wallet/src/utils/state-utils.ts
@@ -1,4 +1,4 @@
-import { ChannelState } from '../redux/states/channels';
+import { ChannelState } from '../redux/channel/state';
 import {
   OWN_COMMITMENT_RECEIVED,
   OPPONENT_COMMITMENT_RECEIVED,
